Start refactoring models
